+++ src/main/java/ru/yandex/stellaburgers/Rest.java
@@ -1,4 +1,4 @@
-package stellaburgers;
+package ru.yandex.stellaburgers;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
