+++ src/main/java/ru/yandex/stellaburgers/UserMethods.java
@@ -1,13 +1,23 @@
-package stellaburgers;
+package ru.yandex.stellaburgers;
 
-public class UserCredentials {
+public class UserMethods {
 
+    private String name;
     private String email;
     private String password;
 
-    public UserCredentials(String email, String password) {
+    public UserMethods(String name, String email, String password) {
+        this.name = name;
         this.email = email;
         this.password = password;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
     }
 
     public String getEmail() {
