+++ src/test/java/ChangeDataUserTests.java
@@ -3,10 +3,10 @@ import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import stellaburgers.User;
-import stellaburgers.UserCredentials;
-import stellaburgers.UserGenerator;
-import stellaburgers.UserMethods;
+import ru.yandex.stellaburgers.User;
+import ru.yandex.stellaburgers.UserCredentials;
+import ru.yandex.stellaburgers.UserGenerator;
+import ru.yandex.stellaburgers.UserMethods;
 import static org.hamcrest.CoreMatchers.equalTo;
 
 public class ChangeDataUserTests {
