+++ src/test/java/CreateOrderTests.java
@@ -3,7 +3,7 @@ import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import stellaburgers.*;
+import ru.yandex.stellaburgers.*;
 import java.util.Arrays;
 import java.util.List;
 import static org.hamcrest.CoreMatchers.equalTo;
